+++ .rokct/cache/marketplace/lib/src/common/presentation/pages/home/home_four/widgets/banner_screen.dart
@@ -31,7 +31,7 @@ import 'package:base_sdk/src/navigation/app_routes.dart';
 
 class BannerScreen extends StatelessWidget {
   final String image;
-  final int bannerId;
+  final String bannerId;
   final String desc;
   final List<ShopData> list;